requirements.txt
asgiref==3.8.1
cffi==1.17.1
colorama==0.4.6
cryptography==45.0.3
Django==5.2.1
django-cors-headers==4.7.0
django-extensions==4.1
django-sslserver==0.22
djangorestframework==3.16.0
MarkupSafe==3.0.2
pillow==11.2.1
pycparser==2.22
pyOpenSSL==25.1.0
python-dotenv==1.1.0
qrcode==8.2
sqlparse==0.5.3
tzdata==2025.2
Werkzeug==3.1.3
